fix(editor): Move project utils
