+++ packages/editor-ui/src/components/CredentialEdit/CredentialSharing.ee.vue
@@ -96,7 +96,7 @@ import type { CredentialScope } from '@n8n/permissions';
 import type { EventBus } from 'n8n-design-system/utils';
 import { useRolesStore } from '@/stores/roles.store';
 import type { RoleMap } from '@/types/roles.types';
-import { ProjectTypes } from '@/features/projects/projects.utils';
+import { ProjectTypes } from '@/utils/projects.utils';
 
 export default defineComponent({
 	name: 'CredentialSharing',
