+++ packages/editor-ui/src/features/projects/components/ProjectCardBadge.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { computed } from 'vue';
 import { useI18n } from '@/composables/useI18n';
-import { ProjectTypes, splitName } from '@/features/projects/projects.utils';
+import { ProjectTypes, splitName } from '@/utils/projects.utils';
 import type { ICredentialsResponse, IWorkflowDb } from '@/Interface';
 import type { Project } from '@/features/projects/projects.types';
