+++ packages/editor-ui/src/features/projects/components/__tests__/projects.utils.test.ts
@@ -1,4 +1,4 @@
-import { splitName } from '@/features/projects/projects.utils';
+import { splitName } from '@/utils/projects.utils';
 
 describe('splitName', () => {
 	test.each([
